[RD-FFK-NET8] Move UI Form in DatesGeneratorUI project plugin fairmat/dev/fairmatserver#3488
